versioning: auto-increment continuous build versions
When we build an artifact from our continuous build job which runs
against changes to `master`, we want to be able to consistently have a
prerelease semantic version, which follows the semantic version ordering
contract. To do this, we can generate and manipulate the version which
is used for the artifact. We use the most recent tag in the current
branch as a base.

As part of this change, we also now expect the publish pipeline to
**not** generate a version; an operator must specify the parameter by
hand.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>